+++ src/VaultSharp/V1/AuthMethods/Kerberos/KerberosAuthMethodInfo.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Net;
-using System.Text;
+﻿using System.Net;
 using VaultSharp.Core;
 
 namespace VaultSharp.V1.AuthMethods.Kerberos
@@ -37,16 +34,27 @@ namespace VaultSharp.V1.AuthMethods.Kerberos
         public ICredentials Credentials { get; }
 
         /// <summary>
+        /// Flag to indicate if the credentials should be cached.
+        /// Defaults to true.
+        /// </summary>
+        /// <value>
+        /// The flag.
+        /// </value>
+        public bool PreAuthenticate { get; }
+
+        /// <summary>
         /// Initializes a new instance of the <see cref="KerberosAuthMethodInfo"/> class.
         /// </summary>
-        public KerberosAuthMethodInfo() : this(AuthMethodType.Kerberos.Type, CredentialCache.DefaultNetworkCredentials)
+        public KerberosAuthMethodInfo() 
+            : this(AuthMethodType.Kerberos.Type, CredentialCache.DefaultCredentials, true)
         {
         }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="KerberosAuthMethodInfo"/> class.
         /// </summary>
-        public KerberosAuthMethodInfo(ICredentials credentials) : this(AuthMethodType.Kerberos.Type, credentials)
+        public KerberosAuthMethodInfo(ICredentials credentials) 
+            : this(AuthMethodType.Kerberos.Type, credentials, true)
         {
         }
 
@@ -56,12 +64,24 @@ namespace VaultSharp.V1.AuthMethods.Kerberos
         /// <param name="mountPoint">The mount point.</param>
         /// <param name="credentials">The credential to use.</param>
         public KerberosAuthMethodInfo(string mountPoint, ICredentials credentials)
+            : this (mountPoint, credentials, true)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="KerberosAuthMethodInfo"/> class.
+        /// </summary>
+        /// <param name="mountPoint">The mount point.</param>
+        /// <param name="credentials">The credential to use.</param>
+        /// <param name="preAuthenticate">The flag to cache credentials.</param>
+        public KerberosAuthMethodInfo(string mountPoint, ICredentials credentials, bool preAuthenticate)
         {
             Checker.NotNull(mountPoint, nameof(mountPoint));
             Checker.NotNull(credentials, nameof(credentials));
 
             MountPoint = mountPoint;
             Credentials = credentials;
+            PreAuthenticate = preAuthenticate;
         }
     }
 }